backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
dj-database-url==2.3.0
psycopg[binary]==3.2.3
django-cors-headers==4.6.0
PyJWT==2.10.1
